requirements.txt
Django==2.0.1
django-allauth==0.34.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
django-rest-auth==0.9.3
django-rest-swagger==2.1.2
dry-rest-permissions==0.1.9
django-import-export==0.7.0
psycopg2==2.7.3.2
django-filter==1.1.0 # Updated from 1.0.4
djangorestframework-gis==0.12
django-richenum==3.4.0
drf-nested-routers==0.90.2
#python-memcached==1.59
#memcached==1.4.33